debugging out-of-bounds clipping
(minor changes to code along the way)
cause: bounding boxes outside of a radius of 1 from the entity
collision is being applied to are not checked, so when traveling at
speeds above 1 block per frame, you can clip into places you're not
supposed to.
fix (not applied yet): finding bboxes for entities needs to account for
entity travel speed.