better vector indexing with v3i_IDX and v3d_IDX
instead of using if/else now I just cast vectors to a pointer and access
it directly. these functions can be used as lvalues too which is pretty
cool. this should definitely have a good impact on performance with some
of the mathier functions

also added DECL_BASIC macros for easier function declaration on my end