ENUMERATION MAGIC.
apparently the C standard specifies that if the first element of an enum
is not given a value, it is given 0, and all subsequent elements not
given values are given the value of the previous element plus one.

this means that you can add an extra element (call it NUM_ELEMENTS or
something) and it will be equal to the number of enumerated values
excluding itself. so I just got to remove a LOT of annoying enum
hackiness :D