variable poisson distribution using pruning
instead of making a fancy capital A Algorithm to produce perlin-noise-y
variable poisson distribution, I created functions to sort and prune
poisson samples using a noise instance. surprisingly clean and
intuitive! I thought this was going to be one of those annoying problems
that takes a week from you out of the blue